Fixed problem that occurs when passing overwrite=True to add_data when the dataset is new and there are NaN values in  the data. Dataset has to be created first before mask is made

+++ cratermaker/components/counting/__init__.py
@@ -463,7 +463,7 @@ class Counting(ComponentBase):
                     d = _convert_tuple_vars(input_dict=d, inverse=False)
                     d = xr.Dataset(data_vars=d).set_coords("id").expand_dims(dim="id")
                     d["id"].attrs["long_name"] = _TALLY_LONG_NAME
-                    d["parent"] = c.id.astype(np.uint32)
+                    d["parent"] = c.id
                     d["parent"].attrs["long_name"] = "id of crater that this ring belongs to"
                     data.append(d)
